resources/network/requirements.txt
networkx
numpy
sklearn
pyyaml
pandas
tensorflow